refactor(api): move badge model to shared
