+++ api/tests/integration/domain/usecases/create-badge_test.js
@@ -4,12 +4,12 @@ import * as badgeRepository from '../../../../lib/infrastructure/repositories/ba
 import * as badgeCriteriaRepository from '../../../../lib/infrastructure/repositories/badge-criteria-repository.js';
 import * as targetProfileRepository from '../../../../lib/infrastructure/repositories/target-profile-repository.js';
 import { createBadge } from '../../../../lib/domain/usecases/create-badge.js';
-import { Badge } from '../../../../lib/domain/models/Badge.js';
 import {
   AlreadyExistingEntityError,
   NotFoundError,
   MissingBadgeCriterionError,
 } from '../../../../lib/domain/errors.js';
+import { Badge } from '../../../../src/shared/domain/models/Badge.js';
 
 describe('Integration | UseCases | create-badge', function () {
   let targetProfileId;
