+++ api/tests/integration/infrastructure/repositories/badge-repository_test.js
@@ -1,6 +1,6 @@
 import { catchErr, databaseBuilder, expect, knex } from '../../../test-helper.js';
 import * as badgeRepository from '../../../../lib/infrastructure/repositories/badge-repository.js';
-import { Badge } from '../../../../lib/domain/models/index.js';
+import { Badge } from '../../../../src/shared/domain/models/Badge.js';
 import lodash from 'lodash';
 import { AlreadyExistingEntityError } from '../../../../lib/domain/errors.js';
